removing duplication code in snippets
