+++ scripts/generateSnippetsIntelliJ.js
@@ -1,19 +1,14 @@
 // To gen snippets in project and in ide
-// node scripts/generateSnippetsAtom.js --genFiles
+// node scripts/generateSnippetsIntelliJ.js --genFiles
 
 // To gen snippets just in ide
-// node scripts/generateSnippetsAtom.js --genFiles --justForIDE
+// node scripts/generateSnippetsIntelliJ.js --genFiles --justForIDE
 
 const fs = require('fs');
 const os = require('os');
 const { argv } = require('yargs');
 
-const givenSteps = require('../uiTestHelpers/stepDefinitions/commonGivenSteps');
-const whenSteps = require('../uiTestHelpers/stepDefinitions/commonWhenSteps');
-const thenSteps = require('../uiTestHelpers/stepDefinitions/commonThenSteps');
-const placeholders = require('../placeholders');
-
-const snippetCodes = {};
+const snippetsCollection = require('./createSnippetsCollection');
 
 let snippets = '<templateSet group="Courgette">\n';
 
@@ -34,51 +29,12 @@ const genSnippet = (matcher, code, varPlaceholders) => {
 `;
 };
 
-const genSnippets = (steps, type) => {
-  if (!snippetCodes[type]) {
-    snippetCodes[type] = [];
-  }
-  steps.forEach((step) => {
-    const allPlaceholders = placeholders.join('|');
-    const varPlaceholders = [];
-    const matcherWithReplacedPlaceholders = step.matcher
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => { varPlaceholders.push(p1); return '\'$var1$\''; })
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => { varPlaceholders.push(p1); return '\'$var2$\''; })
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => { varPlaceholders.push(p1); return '\'$var3$\''; })
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => { varPlaceholders.push(p1); return '\'$var4$\''; })
-      .replace(/\(\?\:(.*)\)\?/g, '$1');
-
-    const zeroOrManyNotMatcher = /\((.*not.*)\)\*/g;
-
-    const newMatcher = step.matcher
-      .replace(/\(\?\:(.*)\)\?/g, (match, p1) => p1.replace(/([^ ]+)/, '_$1_'))
-      .replace(/\(\?\:(.*)\)/g, '$1');
-
-    const generatedCode = `${step.path ?
-      step.path.replace(/[./]*/g, '').replace(/^(actions|checks)/g, '') : 'die'}`;
-
-    const matcher = matcherWithReplacedPlaceholders
-      .replace(zeroOrManyNotMatcher, '')
-      .replace(/\((.*)\)\*/g, '$1')
-      .replace(/\((.*)\)/g, '$1');
-    const newCode = `${type}${step.code || generatedCode}`;
-    snippetCodes[type].push(newCode);
-    genSnippet(matcher, newCode, varPlaceholders);
-
-    if (newMatcher.match(zeroOrManyNotMatcher)) {
-      const newCode2 = `${type}not${step.code || generatedCode}`;
-      const matcher2 = matcherWithReplacedPlaceholders
-        .replace(/\((.*)\)\*/g, '$1');
-      snippetCodes[type].push(newCode2);
-      genSnippet(matcher2, newCode2, varPlaceholders);
-    }
-
-    // console.log(`${type}${step.code || generatedCode}`);
+const genSnippets = () => {
+  snippetsCollection.forEach(({ snippetForXML, code, varPlaceholders }) => {
+    genSnippet(snippetForXML, code, varPlaceholders);
   });
 };
-genSnippets(givenSteps, 'given');
-genSnippets(whenSteps, 'when');
-genSnippets(thenSteps, 'then');
+genSnippets();
 snippets += '</templateSet>';
 
 if (!argv.justForIDE) {
