+++ scripts/generateSnippetsVSCode.js
@@ -8,10 +8,7 @@ const fs = require('fs');
 const os = require('os');
 const { argv } = require('yargs');
 
-const givenSteps = require('../uiTestHelpers/stepDefinitions/commonGivenSteps');
-const whenSteps = require('../uiTestHelpers/stepDefinitions/commonWhenSteps');
-const thenSteps = require('../uiTestHelpers/stepDefinitions/commonThenSteps');
-const placeholders = require('../placeholders');
+const snippetsCollection = require('./createSnippetsCollection');
 
 const ideFolder = `${os.homedir()}/Library/Application Support/Code/User`;
 const ideSnippetsFolder = `${ideFolder}/snippets`;
@@ -39,73 +36,35 @@ const snippetCodes = {};
 
 /* eslint-disable indent */
 
-const genSnippet = (matcher, code) => {
-  const snippet = `{
+const genSnippet = (description, code, snippet) => {
+  const snippetStr = `{
   "Print to console": {
     "prefix": "${code}",
     "body": [
-      "${matcher.replace(/\((.*)\)/g, '$1')}"
+      "${snippet.replace(/\((.*)\)/g, '$1')}"
     ],
-    "description": "${matcher.replace(/\((.*)\)/g, '$1')}"
+    "description": "${description.replace(/\((.*)\)/g, '$1')}"
   }
 }`;
 
   /* eslint-enable indent */
 
   if (!argv.justForIDE) {
-    fs.writeFileSync(`${snippetsFolder}/courgette-${code}.code-snippets`, snippet);
+    fs.writeFileSync(`${snippetsFolder}/courgette-${code}.code-snippets`, snippetStr);
   }
 
   if (ideInstalled) {
-    fs.writeFileSync(`${ideSnippetsFolder}/courgette-${code}.code-snippets`, snippet);
+    fs.writeFileSync(`${ideSnippetsFolder}/courgette-${code}.code-snippets`, snippetStr);
   }
 };
 
-const genSnippets = (steps, type) => {
-  if (!snippetCodes[type]) {
-    snippetCodes[type] = [];
-  }
-  steps.forEach((step) => {
-    const allPlaceholders = placeholders.join('|');
-    const matcherWithReplacedPlaceholders = step.matcher
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => `'\${1:${p1}\}'`)
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => `'\${2:${p1}\}'`)
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => `'\${3:${p1}\}'`)
-      .replace(new RegExp(`'(${allPlaceholders})'`), (m, p1) => `'\${4:${p1}\}'`)
-      .replace(/\(\?\:(.*)\)\?/g, '$1');
-
-    const zeroOrManyNotMatcher = /\((.*not.*)\)\*/g;
-    const newMatcher = step.matcher
-      .replace(/\(\?\:(.*)\)\?/g, (match, p1) =>
-        p1.replace(/([^ ]+)/, '_$1_'))
-      .replace(/\(\?\:(.*)\)/g, '$1');
-
-    const generatedCode = `${step.path ?
-      step.path.replace(/[./]*/g, '').replace(/^(actions|checks)/g, '') : 'die'}`;
-
-    const matcher = matcherWithReplacedPlaceholders
-      .replace(zeroOrManyNotMatcher, '')
-      .replace(/\((.*)\)\*/g, '$1');
-
-    const newCode = `${type}${step.code || generatedCode}`;
-    snippetCodes[type].push(newCode);
-    genSnippet(matcher, newCode);
-
-    if (newMatcher.match(zeroOrManyNotMatcher)) {
-      const newCode2 = `${type}not${step.code || generatedCode}`;
-      const matcher2 = matcherWithReplacedPlaceholders
-        .replace(/\((.*)\)\*/g, '$1');
-      snippetCodes[type].push(newCode2);
-      genSnippet(matcher2, newCode2);
-    }
-
-    // console.log(`${type}${step.code || generatedCode}`);
+const genSnippets = () => {
+  snippetsCollection.forEach(({ description, code, snippet }) => {
+    genSnippet(description, code, snippet);
   });
 };
 
-genSnippets(givenSteps, 'given');
-genSnippets(whenSteps, 'when');
-genSnippets(thenSteps, 'then');
+genSnippets();
 
 if (ideInstalled) {
   console.log(`Snippets added to ${ideSnippetsFolder}`);
